Add files via upload
Handwritten digit recognition is a fundamental task in the field of computer vision, with applications ranging from postal automation to digitizing historical documents. Convolutional Neural Networks (CNNs) have emerged as a powerful tool for this task, offering high accuracy and robustness to variations in handwriting styles.

This project aims to explore and optimize CNN architectures for the task of handwritten digit recognition using the MNIST dataset, a benchmark dataset containing 28x28 pixel grayscale images of handwritten digits from 0 to 9. The project will investigate the impact of various architectural components, such as convolutional layers, pooling layers, and fully connected layers, on the performance of the model.